Add user authentication services and related components
Added UserService for user authentication and operations. Renamed packages to follow proper case convention. Introduced model RegisterBody to handle user registration data. UserRepository was updated to include a method for creating new users. userEntity was also renamed and updated to map to the users table. A userController was created for dealing with users signing up.