Implement search functionality for posts
Added search functionality to fetch posts based on item name or description. The new feature, included in the PostRepository, allows for case insensitive searches. A new Search Service has also been outlined for future improvements. Moreover, a new search function is implemented in the PostService and a corresponding endpoint is available in the PostController.